# packages/shared/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "gagnechris API",
    "version": "0.2.0",
    "description": "HTTP API for Blog CMS and Notebook admin. Same-origin via CloudFront /api/*. Shared DynamoDB single-table (docs/data-model.md)."
  },
  "servers": [
    {
      "url": "https://gagnechris.com",
      "description": "Production"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local API (optional)"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Cognito ID token (Authorization: Bearer <token>)"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "PostStatus": {
        "type": "string",
        "enum": [
          "draft",
          "published",
          "deleted"
        ]
      },
      "PostSeo": {
        "type": "object",
        "nullable": true,
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "ogImage": {
            "type": "string"
          }
        }
      },
...
      "Resume": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1
          },
          "pdfPath": {
            "type": "string",
            "minLength": 1,
            "description": "Site-relative path to the downloadable PDF (publisher always serves /resume.pdf)"
          },
          "content": {
            "$ref": "#/components/schemas/ResumeContent"
          },
          "status": {
            "$ref": "#/components/schemas/PostStatus"
          },
          "publishedAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "seo": {
            "$ref": "#/components/schemas/PostSeo"
          },
          "version": {
            "type": "integer",
            "minimum": 0
          },
          "hasUnpublishedChanges": {
            "type": "boolean",
            "description": "True when a published snapshot exists and the editable draft differs from it"
          }
        },
        "required": [
          "name",
          "pdfPath",
          "content",
          "status",
          "publishedAt",
          "updatedAt",
          "seo",
          "version",
          "hasUnpublishedChanges"
        ]
      },
      "ResumeContent": {
        "type": "object",
        "properties": {
          "summary": {
            "type": "string"
          },
          "competencies": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Flat list; the renderer splits it into two columns"
          },
          "experience": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ResumeExperience"
            }
          },
          "skills": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "education": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ResumeEducation"
            }
          }
        },
        "required": [
          "summary",
          "competencies",
          "experience",
          "skills",
          "education"
        ]
      },
      "ResumeExperience": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "company": {
            "type": "string"
          },
          "bullets": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "title",
          "company",
          "bullets"
        ]
      },
      "ResumeEducation": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "institution": {
            "type": "string"
          },
          "location": {
            "type": "string"
          },
          "year": {
            "type": "string"
          },
          "degreeDetail": {
            "type": "string"
          }
        },
        "required": [
          "title",
          "institution",
          "location",
          "year"
        ]
      },
      "UpdateResumeRequest": {
        "type": "object",
        "properties": {
          "version": {
            "type": "integer",
            "minimum": 0,
            "description": "Expected version for optimistic concurrency"
          },
          "name": {
            "type": "string",
            "minLength": 1
          },
          "pdfPath": {
            "type": "string",
            "minLength": 1
          },
          "content": {
            "$ref": "#/components/schemas/ResumeContent"
          },
          "seo": {
            "$ref": "#/components/schemas/PostSeo"
          }
        },
        "required": [
          "version"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/api/admin/resume": {
      "get": {
        "summary": "Get resume draft (seeded as draft on first read; includes hasUnpublishedChanges)",
        "tags": [
          "Resume"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Resume",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Resume"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update resume (optimistic concurrency via version)",
        "tags": [
          "Resume"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateResumeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Resume"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Version conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/resume/publish": {
      "post": {
        "summary": "Publish resume (copies draft to PUBLISHED snapshot; regenerates PDF)",
        "tags": [
          "Resume"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Published",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Resume"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/resume/unpublish": {
      "post": {
        "summary": "Unpublish resume (live HTML is left in place)",
        "tags": [
          "Resume"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Unpublished (draft)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Resume"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/resume/discard": {
      "post": {
        "summary": "Discard draft edits and restore from the published snapshot",
        "tags": [
          "Resume"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Draft restored from published snapshot",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Resume"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}